changes to func.R
I debugged the multiple factor analysis code, then wrote a getLoadings function. Currently in the midst of a combine loadings function. Have tried many approaches, current best bet looks to be the use of reduce on each of the matrices of loadings. I also realise that matrix algebra is a really useful thing to know more of, and I should resolve to learn more soon